Make FloatingBlockEntity extend BlockLikeEntity
